fix: reorganize services into their own folders
Also renames some things for consistency.
