--- src/durable-objects/index.ts
@@ -1,8 +1,0 @@
-export * from './auth-do';
-export * from './cdn-do';
-export * from './context-do';
-export * from './database-do';
-export * from './scheduler-do';
-export * from './tools-do';
-export * from './image-do';
-export * from './metadata-do';
