+++ src/image-generator/image-generator.do.ts
